+++ legend-sdlc-server-fs/pom.xml
@@ -19,7 +19,7 @@
     <parent>
         <groupId>org.finos.legend.sdlc</groupId>
         <artifactId>legend-sdlc</artifactId>
-        <version>0.138.2-SNAPSHOT</version>
+        <version>0.139.1-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
